+++ src/Zen.Core/SpectreConsole/Integration/IZenCommandGroup.cs
@@ -4,11 +4,11 @@ using System.Linq;
 using System.Threading.Tasks;
 using Spectre.Console.Cli;
 
-namespace Zen.Core.SpectreConsole.Integration
+namespace Zen.Core.SpectreConsole
 {
     public interface IZenCommandGroup
     {
         string Name { get; }
-        void ConfigureCommandApp(in IConfigurator configurator); 
+        void ConfigureCommandApp(IConfigurator<ZenCommandSetting> options); 
     }
 }