+++ src/Zen.Core/SpectreConsole/Integration/ZenCommandSetting.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Threading.Tasks;
 using Spectre.Console.Cli;
 
-namespace Zen.Core.SpectreConsole.Integration
+namespace Zen.Core.SpectreConsole
 {
     public class ZenCommandSetting : CommandSettings
     {